% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelClass.R
\name{gaussianKernel}
\alias{gaussianKernel}
\title{Gaussian Kernel function}
\usage{
gaussianKernel(sigma = 1, l = 1, ...)
}
\arguments{
\item{sigma}{value of sigma of gaussian kernel}

\item{l}{value of lengthscale of gaussian kernel}
}
\value{
a Gram Matrix object from class GramMatrix
}
\description{
It is a function to calculate the Gram matrix from gaussian kernel.
}
\examples{
gaussian_kernel = gaussianKernel(sigma=1, l=1)
z = 1:10
gram_matrix = gaussian_kernel(z,z)
}
