% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelClass.R
\name{get_kpar}
\alias{get_kpar}
\title{Kpar accessor}
\usage{
get_kpar(object)
}
\arguments{
\item{object}{it receives a kernel class}
}
\value{
returns the list of kernel function parameters
}
\description{
Kpar accessor
}
\examples{
gaussian_kernel = gaussianKernel(sigma=1, l=1)
get_kpar(gaussian_kernel)
}
