% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClass.R
\name{graph_test,TestResult-method}
\alias{graph_test,TestResult-method}
\title{Graph method for test result objects}
\usage{
\S4method{graph_test}{TestResult}(object)
}
\arguments{
\item{object}{TestResult object.}
}
\value{
ggplot graph
}
\description{
Graph method for test result objects
}
\examples{
data <- simDataQI(50, dependence_factor=0)
test_result = QuasiIndependenceTest(data)
graph_test(test_result)
}
