% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelClass.R
\docType{class}
\name{Kernel-class}
\alias{Kernel-class}
\title{Kernel class.}
\value{
a kernel class
}
\description{
It is a class that defines a kernel calculation using a function and its parameters.
}
\section{Slots}{

\describe{
\item{\code{function.}}{It receives one function that works as kernel function.}

\item{\code{kpar}}{It is a list with the parameters of the kernel function}
}}

