% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTwoSampleTest.R
\name{KernelTwoSampleTest}
\alias{KernelTwoSampleTest}
\title{A Reproducing-Kernel-Hilbert-Space log-rank test for the two-sample problem}
\usage{
KernelTwoSampleTest(
  data,
  Nboot = 500,
  alpha = 0.05,
  K.fun = "gaussianKernel",
  K.par = list(sigma = 1, l = 1)
)
}
\arguments{
\item{data}{Data must be a data.frame with: X, Delta, Group.}

\item{Nboot}{Number of wildbootstrap repetitions}

\item{alpha}{Significance level}

\item{K.fun}{K matrix kernel function.}

\item{K.par}{List of parameters of K matrix kernel function.}
}
\value{
Returns a Kernel Two Sample Test object.
}
\description{
A Reproducing-Kernel-Hilbert-Space log-rank test for the two-sample problem
}
\examples{
data = simDataTST(100,150)
result = KernelTwoSampleTest(data,Nboot=500,alpha=0.05,K.fun="gaussianKernel",K.par = list("sigma"=1,"l"=1))
}
