% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelClass.R
\name{periodicKernel}
\alias{periodicKernel}
\title{Periodic Kernel function}
\usage{
periodicKernel(sigma = 1, l = 1, p = 1, ...)
}
\arguments{
\item{sigma}{value of sigma of periodic kernel}

\item{l}{value of lengthscale of periodic kernel}

\item{p}{value of period of periodic kernel}
}
\value{
a Gram Matrix object from class GramMatrix
}
\description{
It is a function to calculate the Gram matrix from periodic kernel
}
\examples{
periodic_kernel = periodicKernel(sigma=1, l=1, p = 1)
z = 1:10
gram_matrix = periodic_kernel(z,z)
}
