% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasiIndependenceTest.R
\name{QuasiIndependenceTest}
\alias{QuasiIndependenceTest}
\title{Quasi Independence Test}
\usage{
QuasiIndependenceTest(
  data,
  Nboot = 500,
  alpha = 0.05,
  K.fun = "gaussianKernel",
  L.fun = "gaussianKernel",
  K.par = list(sigma = 1, l = 1),
  L.par = list(sigma = 1, l = 1)
)
}
\arguments{
\item{data}{Data must be a data.frame with: X.ob,T.ob,Delta.}

\item{Nboot}{WildBoostrap sample size.}

\item{alpha}{Significance level}

\item{K.fun}{K matrix kernel function.}

\item{L.fun}{L matrix kernel function.}

\item{K.par}{List of parameters of K matrix kernel function.}

\item{L.par}{List of parameters of L matrix kernel function.}
}
\value{
Returns a Quasi Independence Test object.
}
\description{
This test was developed to determine if it exists an association between X and Y in the truncated setting.
}
\examples{
data <-  simDataQI(n=50, dependence_factor=0.65)

QuasiIndependenceTest(data,Nboot=500,alpha=0.05,K.fun="gaussianKernel",L.fun="gaussianKernel",K.par=list("sigma"=1,"l"=1),L.par=list("sigma"=1,"l"=1))
}
