% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClass.R
\name{show,QuasiIndependenceTest-method}
\alias{show,QuasiIndependenceTest-method}
\title{Show method for Quasi Independence Test}
\usage{
\S4method{show}{QuasiIndependenceTest}(object)
}
\arguments{
\item{object}{Quasi Independence Test object.}
}
\value{
None
}
\description{
Show method for Quasi Independence Test
}
\examples{
data <- simDataQI(50, dependence_factor=0)
test_result = QuasiIndependenceTest(data)
show(test_result)
}
