% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClass.R
\name{show,TestResult-method}
\alias{show,TestResult-method}
\title{show method for test result}
\usage{
\S4method{show}{TestResult}(object)
}
\arguments{
\item{object}{TestResult class.}
}
\value{
None
}
\description{
show method for test result
}
