% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClass.R
\name{show,TwoSampleTest-method}
\alias{show,TwoSampleTest-method}
\title{Show method for Two Sample Test}
\usage{
\S4method{show}{TwoSampleTest}(object)
}
\arguments{
\item{object}{Two Sample Test object.}
}
\value{
None
}
\description{
Show method for Two Sample Test
}
\examples{
data <- simDataTST(50,50)
test_result = KernelTwoSampleTest(data)
show(test_result)
}
