% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdataTST.R
\name{simDataTST}
\alias{simDataTST}
\title{#' Simulated Data Function for Two Sample Test}
\usage{
simDataTST(n1 = 100, n2 = 100, rate = list(t1 = 3, c1 = 1, t2 = 0.5, c2 = 0.1))
}
\arguments{
\item{n1}{Number of observations of group 1}

\item{n2}{Number of observations of group 2}

\item{rate}{A list with four values: rates for t1, c1, t2, c2}
}
\value{
A dataframe with 3 values: X, Delta and Group
}
\description{
#' Simulated Data Function for Two Sample Test
}
\examples{
data = simDataTST(100,100)
}
