% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClass.R
\docType{class}
\name{TestResult-class}
\alias{TestResult-class}
\title{Test Result class}
\value{
a test result class object
}
\description{
Test Result class
}
\section{Slots}{

\describe{
\item{\code{t.stat}}{value of stadistic.}

\item{\code{w.boot}}{vector of wild bootstrap values.}

\item{\code{result}}{result of the test.}

\item{\code{p.value}}{p value of the test.}

\item{\code{quantile}}{quantile of the test.}
}}

